Move SpecConfigData
