+++ data/provider/src/main/java/tech/pegasys/teku/api/ConfigProvider.java
@@ -18,7 +18,6 @@ import org.apache.tuweni.units.bigints.UInt256;
 import tech.pegasys.teku.api.response.v1.config.GetForkScheduleResponse;
 import tech.pegasys.teku.api.response.v1.config.GetSpecResponse;
 import tech.pegasys.teku.api.schema.Fork;
-import tech.pegasys.teku.ethereum.json.types.config.SpecConfigData;
 import tech.pegasys.teku.infrastructure.unsigned.UInt64;
 import tech.pegasys.teku.spec.Spec;
 import tech.pegasys.teku.spec.config.SpecConfig;
