+++ ethereum/json-types/build.gradle
@@ -4,7 +4,6 @@ dependencies {
     implementation project(':data:serializer')
     implementation project(':infrastructure:restapi')
     implementation project(':infrastructure:http')
-    implementation project(':data:provider')
 
     testImplementation testFixtures(project(':ethereum:spec'))
 }
